run prettier on js-files and remove linting
i ran `npx prettier --write` on the files in this commit, so they would
be easier to work with. we might want to do this with the markdown-files
at some point in time and also automate this later on.

also eslint has been removed as we already had a whopping 1033 problems
with 876 errors. it didn't seem worthwhile to fix that when there is
something like prettier.js doing this in a matter of seconds.